Mutate a (factor, character, integer, etc) column into multiple columns,

Easily mutate a single column into multiple columns (~dummies+1), while retaining variable labels and order of the original factor variable.

Usage

col_to_binaries(data, col, var_separator = "___", label_separator = " - ")

Arguments

data

Survey data

obj:<data.frame>|obj:<tbl_df> // Required

A data frame (or a srvyr-object) with the columns specified in the chapter_overview 'dep_cat', etc columns.

col

Single column. Tidy-select.

var_separator

Variable separator

scalar<character> // default: NULL (optional)

Separator between old variable name and categories.

label_separator

scalar<character> // default: NULL (optional)

Separator between old label name and new label part.

Value

Original data frame with the binary columns attached, containing new labels.

Examples

col_to_binaries(ex_survey, col = b_3, label_separator = "  -  ")
#> # A tibble: 300 × 35
#>    x1_sex  x2_human     x3_nationality a_1   a_2   a_3   a_4   a_5   a_6   a_7  
#>    <fct>   <fct>        <fct>          <fct> <fct> <fct> <fct> <fct> <fct> <fct>
#>  1 Males   Definitely … Eswatini       Yes   Yes   Yes   No    Yes   Yes   Yes  
#>  2 Females Robot?       Denmark        NA    Yes   No    Yes   Yes   No    No   
#>  3 Females Robot?       Eswatini       No    No    No    No    No    No    Yes  
#>  4 Males   Robot?       Cameroon       Yes   No    Yes   No    No    No    No   
#>  5 Males   Definitely … Andorra        No    Yes   No    No    No    Yes   No   
#>  6 Males   Robot?       Ghana          Yes   Yes   Yes   No    Yes   Yes   Yes  
#>  7 Males   Definitely … Belize         Yes   No    No    Yes   Yes   NA    No   
#>  8 Females Definitely … Cameroon       No    Yes   Yes   Yes   Yes   No    No   
#>  9 Females Robot?       Denmark        No    NA    No    No    No    NA    No   
#> 10 Females Definitely … Denmark        Yes   Yes   No    No    NA    No    No   
#> # ℹ 290 more rows
#> # ℹ 25 more variables: a_8 <fct>, a_9 <fct>, b_1 <fct>, b_2 <fct>, b_3 <fct>,
#> #   c_1 <dbl>, c_2 <dbl>, d_1 <fct>, d_2 <fct>, d_3 <fct>, d_4 <fct>,
#> #   e_1 <fct>, e_2 <fct>, e_3 <fct>, e_4 <fct>, p_1 <fct>, p_2 <fct>,
#> #   p_3 <fct>, p_4 <fct>, f_uni <chr>, open_comments <chr>, resp_status <int>,
#> #   `b_3___Not at all` <int>, `b_3___A bit` <int>, `b_3___A lot` <int>