Title

Title

Usage

embed_uni_sigtest(
  data,
  dep,
  ...,
  indep = NULL,
  .variable_type = NULL,
  indep_type = NULL,
  mesos_group = NULL,
  call = rlang::caller_env()
)

Arguments

data

Survey data

obj:<data.frame>|obj:<tbl_df> // Required

A data frame (or a srvyr-object) with the columns specified in the chapter_overview 'dep_cat', etc columns.

dep, indep

Variable selections

<tidyselect> // Default: NULL, meaning everything for dep, nothing for indep.

Columns in data. Currently allows tidyselect-syntax, which will be removed.

...

Dynamic dots

<dynamic-dots>

Arguments forwarded to the corresponding functions that create the elements.

.variable_type, indep_type

Optional string specifying data type ("int", "fct", "character", "factor", "numeric", "integer"). Can be obtained from data lookup.

mesos_group

Specific group to compare with

scalar<character> // Default: NULL (optional)

Both the absolute and relative folderpaths are required, as strings.

call

Internal call

obj:<call> // Default: rlang::caller_env() (optional)

Both the absolute and relative folderpaths are required, as strings.

Value

Data frame

Examples

embed_uni_sigtest(data=ex_survey1,
                  dep = a_1:a_9, reps=10, hide_test_if_n_below=10)
#>   variable .variable_label test stat   p
#> 1      a_1    Agreement #1 prop  0.5 0.8
#> 2      a_2    Agreement #2 prop  0.7 0.3
#> 3      a_3    Agreement #3 prop  0.5 1.0
#> 4      a_4    Agreement #4 prop  0.4 0.6
#> 5      a_5    Agreement #5 prop  0.3 0.3
#> 6      a_6    Agreement #6 prop  0.6 0.2
#> 7      a_7    Agreement #7 prop  0.5 1.0
#> 8      a_8    Agreement #8 prop  0.5 0.8
#> 9      a_9    Agreement #9 prop  0.3 0.2
embed_uni_sigtest(data=ex_survey1,
                  dep = b_1:b_3, reps=10, hide_test_if_n_below=10,
                  label_separator=" - ")
#>   variable .variable_label  test stat   p
#> 1      b_1          Bejing  <NA>   NA  NA
#> 2      b_2        Brussels chisq 21.4 0.3
#> 3      b_3        Budapest chisq 18.7 0.3
embed_uni_sigtest(data=ex_survey1,
                  dep = c_1:c_2, reps=10, hide_test_if_n_below=10)
#>   variable .variable_label test stat   p
#> 1      c_1       Company A mean 18.8 0.3
#> 2      c_2       Company B mean 20.2 0.3