Generate Quarto Index File That Merges All Chapters

This function creates an index Quarto file (QMD) that merges all chapters in the specified order. It can also include optional title and author(s) information.

Usage

gen_qmd_index(
  index_filepath = "complete_report.qmd",
  ...,
  chapter_filepaths = NULL,
  call = rlang::caller_env()
)

Arguments

index_filepath

Path to the output index.qmd file. Defaults to "complete_report.qmd"

...

Dynamic dots

<dynamic-dots>

Arguments forwarded to the corresponding functions that create the elements.

chapter_filepaths

A character vector containing the filepaths of the chapter files to be included in the index, in the desired order.

call

Internal call argument. Not to be fiddled with by the user.

Value

A string containing the filepath to the generated Quarto index file.

Examples

filepath <- tempfile(pattern = "index", fileext = ".qmd")
index_file <-
  saros:::gen_qmd_index(
                 authors = c("Marky Twain",
                             "J.K. Rawlings",
                             "Stephen Prince"),
               chapter_filepaths =
                 c("Ch1.qmd",
                 "Ch2.qmd",
                 "Ch3.qmd"),
                 index_filepath = filepath)