Read Default Arguments for draft_report() from YAML-file

Read Default Arguments for draft_report() from YAML-file

Usage

read_default_draft_report_args(path)

Arguments

path

scalar<character> // Required. default: settings.yaml

Value

The defaults as a yaml-object.

Examples

path <- write_default_draft_report_args(path=tempfile(fileext=".yaml"))
read_default_draft_report_args(path=path)
#> $title
#> [1] "Report"
#> 
#> $authors
#> NULL
#> 
#> $mesos_report
#> [1] FALSE
#> 
#> $mesos_var
#> NULL
#> 
#> $label_separator
#> [1] " - "
#> 
#> $name_separator
#> NULL
#> 
#> $chapter_yaml_file
#> NULL
#> 
#> $chapter_qmd_start_section_filepath
#> NULL
#> 
#> $chapter_qmd_end_section_filepath
#> NULL
#> 
#> $index_filename
#> [1] "index"
#> 
#> $index_yaml_file
#> NULL
#> 
#> $index_qmd_start_section_filepath
#> NULL
#> 
#> $index_qmd_end_section_filepath
#> NULL
#> 
#> $report_filename
#> [1] "0_report"
#> 
#> $report_yaml_file
#> NULL
#> 
#> $report_qmd_start_section_filepath
#> NULL
#> 
#> $report_qmd_end_section_filepath
#> NULL
#> 
#> $element_names
#>  [1] "uni_cat_prop_plot"    "uni_cat_freq_plot"    "uni_cat_table"       
#>  [4] "uni_chr_table"        "hline"                "bi_catcat_prop_plot" 
#>  [7] "bi_catcat_freq_plot"  "bi_catcat_prop_plot2" "bi_catcat_freq_plot2"
#> [10] "bi_catcat_table"      "bi_sigtest"          
#> 
#> $sort_by
#> [1] ".upper"
#> 
#> $data_label
#> [1] "proportion"      "percentage"      "percentage_bare" "count"          
#> [5] "mean"            "median"         
#> 
#> $always_show_bi_for_indep
#> NULL
#> 
#> $categories_treated_as_na
#> NULL
#> 
#> $variables_always_at_top
#> NULL
#> 
#> $variables_always_at_bottom
#> NULL
#> 
#> $return_raw
#> [1] TRUE
#> 
#> $showNA
#> [1] "never"  "always" "ifany" 
#> 
#> $totals
#> [1] FALSE
#> 
#> $hide_bi_entry_if_sig_above
#> [1] 1
#> 
#> $hide_test_if_n_below
#> [1] 10
#> 
#> $hide_result_if_n_below
#> [1] 10
#> 
#> $hide_chr_for_others
#> [1] TRUE
#> 
#> $hide_variable_if_all_na
#> [1] TRUE
#> 
#> $single_y_bivariates_if_indep_cats_above
#> [1] 3
#> 
#> $single_y_bivariates_if_deps_above
#> [1] 20
#> 
#> $digits
#> [1] 1
#> 
#> $data_label_decimal_symbol
#> [1] "."
#> 
#> $hide_label_if_prop_below
#> [1] 0.01
#> 
#> $hide_axis_text_if_single_variable
#> [1] FALSE
#> 
#> $main_font_size
#> [1] 10
#> 
#> $label_font_size
#> [1] 3
#> 
#> $strip_font_size
#> [1] 7
#> 
#> $legend_font_size
#> [1] 7
#> 
#> $strip_width
#> [1] 15
#> 
#> $strip_angle
#> [1] 0
#> 
#> $x_axis_label_width
#> [1] 20
#> 
#> $plot_height_multiplier_per_horizontal_line
#> [1] NA
#> 
#> $plot_height_multiplier_per_vertical_letter
#> [1] 0.2
#> 
#> $plot_height_multiplier_per_facet
#> [1] 0.95
#> 
#> $plot_height_multiplier_per_legend_line
#> [1] 1.1
#> 
#> $plot_height_fixed_constant
#> [1] 0
#> 
#> $plot_height_max
#> [1] 8
#> 
#> $plot_height_min
#> [1] 1.5
#> 
#> $vertical_height
#> [1] 12
#> 
#> $vertical
#> [1] FALSE
#> 
#> $png_scale
#> [1] 1.2
#> 
#> $png_width
#> [1] 14
#> 
#> $png_height
#> [1] 16
#> 
#> $font_family
#> [1] "sans"
#> 
#> $colour_palette_nominal
#> NULL
#> 
#> $colour_palette_ordinal
#> NULL
#> 
#> $colour_na
#> [1] "gray90"
#> 
#> $colour_2nd_binary_cat
#> NULL
#> 
#> $table_main_question_as_header
#> [1] FALSE
#> 
#> $max_width_obj
#> [1] 128
#> 
#> $max_width_file
#> [1] 64
#> 
#> $max_clean_folder_name
#> [1] 12
#> 
#> $max_path_warning_threshold
#> [1] 260
#> 
#> $open_after_drafting
#> [1] FALSE
#> 
#> $organize_by
#> [1] "chapter"                    ".variable_label_prefix_dep"
#> [3] ".variable_name_indep"       ".element_name"             
#> 
#> $arrange_section_by
#> [1] FALSE FALSE FALSE FALSE
#> 
#> $na_first_in_section
#> [1] TRUE
#> 
#> $ignore_heading_for_group
#> [1] ".element_name"        ".variable_type_dep"   ".variable_type_indep"
#> [4] "chapter"             
#> 
#> $replace_heading_for_group
#> [1] ".variable_name_dep"   ".variable_name_indep"
#> 
#> $mesos_first
#> [1] TRUE
#> 
#> $descend
#> [1] TRUE
#> 
#> $require_common_categories
#> [1] TRUE
#> 
#> $panel_tabset_mesos
#> [1] TRUE
#> 
#> $pdf
#> [1] TRUE
#> 
#> $attach_chapter_dataset
#> [1] TRUE
#> 
#> $auxiliary_variables
#> NULL
#> 
#> $micro
#> [1] FALSE
#> 
#> $reps
#> [1] 1000
#> 
#> $information
#>  [1] ".variable_label_dep" ".category"           ".count"             
#>  [4] ".count_se"           ".proportion"         ".proportion_se"     
#>  [7] ".mean"               ".mean_se"            ".data_label"        
#> [10] ".comb_categories"    ".sum_value"         
#> 
#> $contents
#>  [1] "intro"             "not_used_category" "mode_max"         
#>  [4] "value_max"         "value_min"         "value_diff"       
#>  [7] "mean_max"          "mean_min"          "mean_diff"        
#> [10] "median_max"        "median_min"        "median_diff"      
#> [13] "variance_max"      "variance_min"     
#> 
#> $include_numbers
#> [1] TRUE
#> 
#> $n_top_bottom
#> [1] 1
#> 
#> $log_file
#> NULL
#> 
#> $serialized_format
#> [1] "rds" "qs" 
#> 
#> $tabular_format
#> [1] "delim" "xlsx"  "csv"   "csv2"  "tsv"   "sav"   "dta"  
#> 
#> $translations
#>                                                                  last_sep 
#>                                                                   " and " 
#>                                                              intro_prefix 
#>                      "We will now look at the questions asked regarding " 
#>                                                              intro_suffix 
#>                                                                        "" 
#>                                                            mode_max_onfix 
#>                                                                    " on " 
#>                                                           mode_max_prefix 
#>                                         "The most common responses were " 
#>                                                           mode_max_suffix 
#>                                                                        "" 
#>                                                           not_used_prefix 
#>                       "The following response categories were not used: " 
#>                                                           not_used_suffix 
#>                                                                        "" 
#>                                                          value_max_prefix 
#>                                                                        "" 
#>                                                           value_max_infix 
#>   " {?is/are} the {dots$n_top_bottom} item{?s} where the most responded " 
#>                                                          value_max_suffix 
#>                                                                        "" 
#>                                                          value_min_prefix 
#>                                                                        "" 
#>                                                           value_min_infix 
#> " {?is/are} the {dots$n_top_bottom} item{?s} where the fewest responded " 
#>                                                          value_min_suffix 
#>                                                                        "" 
#>                                                                mean_onfix 
#>                                                                    "M = " 
#>                                                           mean_max_prefix 
#>                                              "They have highest mean on " 
#>                                                           mean_max_suffix 
#>                                                                        "" 
#>                                                           mean_min_prefix 
#>                                               "They have lowest mean on " 
#>                                                           mean_min_suffix 
#>                                                                        "" 
#>                                                              median_onfix 
#>                                                               "Median = " 
#>                                                         median_max_prefix 
#>                                            "They have highest median on " 
#>                                                         median_max_suffix 
#>                                                                        "" 
#>                                                         median_min_prefix 
#>                                             "They have lowest median on " 
#>                                                         median_min_suffix 
#>                                                                        "" 
#>                                                           intro_by_prefix 
#>                      "We will now look at the questions asked regarding " 
#>                                                            intro_by_infix 
#>                                                        " broken down by " 
#>                                                           intro_by_suffix 
#>                                                                        "" 
#>                                                              by_breakdown 
#>                                                                    " by " 
#>                                                            n_equal_prefix 
#>                                                           " (N &equals; " 
#>                                                            n_equal_suffix 
#>                                                                       ")" 
#>                                                           table_heading_N 
#>                                                               "Total (N)" 
#>                                                                  by_total 
#>                                                                "Everyone" 
#>                                                            sigtest_prefix 
#>                                                "Significance testing of " 
#>                                                            sigtest_suffix 
#>                                                                        "" 
#>                                                        mesos_group_prefix 
#>                                                                " Group: " 
#>                                                        mesos_group_suffix 
#>                                                                        "" 
#>                                                    mesos_label_all_others 
#>                                                                  "Others" 
#>