Processes A 'chapter_overview' Data Frame

Processes A 'chapter_overview' Data Frame

Usage

refine_chapter_overview(
  chapter_overview = NULL,
  data = NULL,
  ...,
  progress = TRUE,
  variable_group_dep = ".variable_group_dep",
  call = rlang::caller_env()
)

Arguments

chapter_overview

What goes in each chapter

obj:<data.frame>|obj:<tbl_df> // Required

Data frame (or tibble, possibly grouped). One row per chapter. Should contain the columns 'chapter' and 'dep', Optionally 'indep' (independent variables) and other informative columns as needed.

data

Survey data

obj:<data.frame>|obj:<tbl_df> // Required

A data frame (or a srvyr-object) with the columns specified in the chapter_overview 'dep_cat', etc columns.

...

Dynamic dots

<dynamic-dots>

Arguments forwarded to the corresponding functions that create the elements.

progress

Whether to display progress message

scalar<logical> // default: TRUE

variable_group_dep

Name for the variable_group_dep column

scalar<string> // default: ".variable_group_dep"

This column is used to group variables that are part of the same bivariate analysis.

call

Internal call

obj:<call> // Default: rlang::caller_env() (optional)

Both the absolute and relative folderpaths are required, as strings.

Value

Grouped tibble.

Examples

ref_df <- refine_chapter_overview(chapter_overview = ex_survey_ch_overview)
#> Refining chapter_overview...
# \donttest{
ref_df2 <- refine_chapter_overview(chapter_overview = ex_survey_ch_overview,
                     data = ex_survey, hide_bi_entry_if_sig_above=.05)
#> Refining chapter_overview...
#> Removing bivariate occurences if `hide_bi_entry_if_sig_above`: `0.05`, except .
#> Refining chapter_overview...
#> Not using the following variables in `data`: `x1_sex`, `x2_human`, `a_7`,
#> `a_8`, `f_uni`, and `resp_status`.
# }