Read Default Arguments for draft_report() from YAML-file

Read Default Arguments for draft_report() from YAML-file

Usage

read_default_draft_report_args(path)

Arguments

path

scalar<character> // Required. default: settings.yaml

Value

The defaults as a yaml-object.

Examples

path <- write_default_draft_report_args(path=tempfile(fileext=".yaml"))
read_default_draft_report_args(path=path)
#> $title
#> [1] "Report"
#> 
#> $authors
#> NULL
#> 
#> $mesos_report
#> [1] FALSE
#> 
#> $mesos_var
#> NULL
#> 
#> $label_separator
#> [1] " - "
#> 
#> $name_separator
#> NULL
#> 
#> $index_yaml_file
#> NULL
#> 
#> $report_yaml_file
#> NULL
#> 
#> $chapter_yaml_file
#> NULL
#> 
#> $qmd_start_section_filepath
#> NULL
#> 
#> $qmd_end_section_filepath
#> NULL
#> 
#> $index_filename
#> [1] "index.qmd"
#> 
#> $element_names
#>  [1] "uni_cat_prop_plot"    "uni_cat_freq_plot"    "uni_cat_table"       
#>  [4] "uni_chr_table"        "hline"                "bi_catcat_prop_plot" 
#>  [7] "bi_catcat_freq_plot"  "bi_catcat_prop_plot2" "bi_catcat_freq_plot2"
#> [10] "bi_catcat_table"      "bi_sigtest"          
#> 
#> $sort_by
#> [1] ".upper"
#> 
#> $data_label
#> [1] "proportion"      "percentage"      "percentage_bare" "count"          
#> [5] "mean"            "median"         
#> 
#> $always_show_bi_for_indep
#> NULL
#> 
#> $categories_treated_as_na
#> NULL
#> 
#> $variables_always_at_top
#> NULL
#> 
#> $variables_always_at_bottom
#> NULL
#> 
#> $return_raw
#> [1] TRUE
#> 
#> $showNA
#> [1] "never"  "always" "ifany" 
#> 
#> $totals
#> [1] FALSE
#> 
#> $hide_bi_entry_if_sig_above
#> [1] 1
#> 
#> $hide_test_if_n_below
#> [1] 10
#> 
#> $hide_result_if_n_below
#> [1] 10
#> 
#> $hide_chr_for_others
#> [1] TRUE
#> 
#> $hide_variable_if_all_na
#> [1] TRUE
#> 
#> $single_y_bivariates_if_indep_cats_above
#> [1] 3
#> 
#> $single_y_bivariates_if_deps_above
#> [1] 20
#> 
#> $digits
#> [1] 1
#> 
#> $data_label_decimal_symbol
#> [1] "."
#> 
#> $hide_label_if_prop_below
#> [1] 0.01
#> 
#> $hide_axis_text_if_single_variable
#> [1] FALSE
#> 
#> $main_font_size
#> [1] 10
#> 
#> $label_font_size
#> [1] 3
#> 
#> $strip_font_size
#> [1] 7
#> 
#> $legend_font_size
#> [1] 7
#> 
#> $strip_width
#> [1] 15
#> 
#> $strip_angle
#> [1] 0
#> 
#> $x_axis_label_width
#> [1] 20
#> 
#> $plot_height_multiplier_per_horizontal_line
#> [1] NA
#> 
#> $plot_height_multiplier_per_vertical_letter
#> [1] 0.2
#> 
#> $plot_height_multiplier_per_facet
#> [1] 0.95
#> 
#> $plot_height_multiplier_per_legend_line
#> [1] 1.1
#> 
#> $plot_height_fixed_constant
#> [1] 0
#> 
#> $plot_height_max
#> [1] 8
#> 
#> $plot_height_min
#> [1] 1.5
#> 
#> $vertical_height
#> [1] 12
#> 
#> $vertical
#> [1] FALSE
#> 
#> $png_scale
#> [1] 1.2
#> 
#> $png_width
#> [1] 14
#> 
#> $png_height
#> [1] 16
#> 
#> $font_family
#> [1] "sans"
#> 
#> $colour_palette_nominal
#> NULL
#> 
#> $colour_palette_ordinal
#> NULL
#> 
#> $colour_na
#> [1] "gray90"
#> 
#> $colour_2nd_binary_cat
#> NULL
#> 
#> $table_main_question_as_header
#> [1] FALSE
#> 
#> $max_width_obj
#> [1] 128
#> 
#> $max_width_file
#> [1] 64
#> 
#> $max_clean_folder_name
#> [1] 12
#> 
#> $open_after_drafting
#> [1] FALSE
#> 
#> $organize_by
#> [1] "chapter"                    ".variable_label_prefix_dep"
#> [3] ".variable_name_indep"       ".element_name"             
#> 
#> $arrange_output_by
#> [1] "chapter"              ".variable_name_dep"   ".variable_name_indep"
#> 
#> $ignore_heading_for_group
#> [1] ".element_name"      ".variable_type_dep" "chapter"           
#> 
#> $replace_heading_for_group
#> [1] ".variable_name_dep"
#> 
#> $mesos_first
#> [1] TRUE
#> 
#> $descend
#> [1] TRUE
#> 
#> $require_common_categories
#> [1] TRUE
#> 
#> $panel_tabset_mesos
#> [1] TRUE
#> 
#> $pdf
#> [1] TRUE
#> 
#> $attach_chapter_dataset
#> [1] TRUE
#> 
#> $auxiliary_variables
#> NULL
#> 
#> $flexi
#> [1] FALSE
#> 
#> $micro
#> [1] FALSE
#> 
#> $reps
#> [1] 1000
#> 
#> $information
#>  [1] ".variable_label_dep" ".category"           ".count"             
#>  [4] ".count_se"           ".proportion"         ".proportion_se"     
#>  [7] ".mean"               ".mean_se"            ".data_label"        
#> [10] ".comb_categories"    ".sum_value"         
#> 
#> $contents
#>  [1] "intro"             "not_used_category" "mode_max"         
#>  [4] "value_max"         "value_min"         "value_diff"       
#>  [7] "mean_max"          "mean_min"          "mean_diff"        
#> [10] "median_max"        "median_min"        "median_diff"      
#> [13] "variance_max"      "variance_min"     
#> 
#> $include_numbers
#> [1] TRUE
#> 
#> $n_top_bottom
#> [1] 1
#> 
#> $log_file
#> NULL
#> 
#> $serialized_format
#> [1] "rds" "qs" 
#> 
#> $tabular_format
#> [1] "delim" "xlsx"  "csv"   "csv2"  "tsv"   "sav"   "dta"  
#> 
#> $translations
#>                                                                  last_sep 
#>                                                                   " and " 
#>                                                           download_report 
#>                                                   "Download report (PDF)" 
#>                                                              intro_prefix 
#>                      "We will now look at the questions asked regarding " 
#>                                                              intro_suffix 
#>                                                                        "" 
#>                                                            mode_max_onfix 
#>                                                                    " on " 
#>                                                           mode_max_prefix 
#>                                         "The most common responses were " 
#>                                                           mode_max_suffix 
#>                                                                        "" 
#>                                                           not_used_prefix 
#>                       "The following response categories were not used: " 
#>                                                           not_used_suffix 
#>                                                                        "" 
#>                                                          value_max_prefix 
#>                                                                        "" 
#>                                                           value_max_infix 
#>   " {?is/are} the {dots$n_top_bottom} item{?s} where the most responded " 
#>                                                          value_max_suffix 
#>                                                                        "" 
#>                                                          value_min_prefix 
#>                                                                        "" 
#>                                                           value_min_infix 
#> " {?is/are} the {dots$n_top_bottom} item{?s} where the fewest responded " 
#>                                                          value_min_suffix 
#>                                                                        "" 
#>                                                                mean_onfix 
#>                                                                    "M = " 
#>                                                           mean_max_prefix 
#>                                              "They have highest mean on " 
#>                                                           mean_max_suffix 
#>                                                                        "" 
#>                                                           mean_min_prefix 
#>                                               "They have lowest mean on " 
#>                                                           mean_min_suffix 
#>                                                                        "" 
#>                                                              median_onfix 
#>                                                               "Median = " 
#>                                                         median_max_prefix 
#>                                            "They have highest median on " 
#>                                                         median_max_suffix 
#>                                                                        "" 
#>                                                         median_min_prefix 
#>                                             "They have lowest median on " 
#>                                                         median_min_suffix 
#>                                                                        "" 
#>                                                           intro_by_prefix 
#>                      "We will now look at the questions asked regarding " 
#>                                                            intro_by_infix 
#>                                                        " broken down by " 
#>                                                           intro_by_suffix 
#>                                                                        "" 
#>                                                              by_breakdown 
#>                                                                    " by " 
#>                                                            n_equal_prefix 
#>                                                           " (N &equals; " 
#>                                                            n_equal_suffix 
#>                                                                       ")" 
#>                                                           table_heading_N 
#>                                                               "Total (N)" 
#>                                                                  by_total 
#>                                                                "Everyone" 
#>                                                            sigtest_prefix 
#>                                                "Significance testing of " 
#>                                                            sigtest_suffix 
#>                                                                        "" 
#>                                                        mesos_group_prefix 
#>                                                                " Group: " 
#>                                                        mesos_group_suffix 
#>                                                                        "" 
#>                                                    mesos_label_all_others 
#>                                                                  "Others" 
#>                                                          empty_chunk_text 
#>                                                                "\nText\n" 
#>                                                       flexi_input_chapter 
#>                                                             "Chapter(s):" 
#>                                                           flexi_input_dep 
#>                                                  "Dependent variable(s):" 
#>                                                         flexi_input_indep 
#>                                                   "Independent variable:" 
#>                                                   flexi_input_mesos_group 
#>                                                                 "Filter:" 
#>                                                         flexi_figure_type 
#>                                                            "Figure type:" 
#>                                                          flexi_data_label 
#>                                                      "Summary to display" 
#>                                                              flexi_showNA 
#>                                                       "Show NA (Missing)" 
#>                                                             flexi_sort_by 
#>                                                                 "Sort by" 
#>                                                              flexi_totals 
#>                                                                  "Totals" 
#>                                                              flexi_digits 
#>                                                    "Digits after decimal" 
#>                                                               flexi_table 
#>                                                                   "Table" 
#>                                                              flexi_figure 
#>                                                                  "Figure" 
#>                                                  flexi_cols_variable_name 
#>                                                           "Variable name" 
#>                                                 flexi_cols_variable_label 
#>                                                          "Variable label" 
#>                                                       flexi_cols_category 
#>                                                       "Response category" 
#>                                                          flexi_cols_count 
#>                                                                       "N" 
#>                                                       flexi_cols_count_se 
#>                                                                   "SE(N)" 
#>                                                     flexi_cols_proportion 
#>                                                              "Proportion" 
#>                                                  flexi_cols_proportion_se 
#>                                                          "SE(Proportion)" 
#>                                                           flexi_cols_mean 
#>                                                                    "Mean" 
#>                                                        flexi_cols_mean_se 
#>                                                                "SE(Mean)" 
#>                                                     flexi_cols_data_label 
#>                                                              "Data label" 
#>                                                flexi_cols_comb_categories 
#>                                                     "Combined categories" 
#>                                                      flexi_cols_sum_value 
#>                            "Sum of data label across combined categories" 
#>                                                            flexi_validate 
#>                     "Error: Columns must have some categories in common." 
#>                                                            flexi_settings 
#>                                                                "Settings" 
#>                                                      flexi_basic_settings 
#>                                                                   "Basic" 
#>                                                   flexi_advanced_settings 
#>                                                                "Advanced" 
#>                                                    flexi_input_indep_none 
#>                                                                  "<none>" 
#>                                              flexi_figure_type_proportion 
#>                                                              "Proportion" 
#>                                               flexi_figure_type_frequency 
#>                                                               "Frequency" 
#>                                            flexi_hide_label_if_prop_below 
#>                                         "Hide label if proportion below:" 
#>