saros: Semi-Automatic Reporting of Ordinary Surveys

Offers a systematic way for conditional reporting of figures and tables for many (and bivariate combinations of) variables, typically from survey data. Contains interactive 'ggiraph'-based (https://CRAN.R-project.org/package=ggiraph) plotting functions and data frame-based summary tables (bivariate significance tests, frequencies/proportions, unique open ended responses, etc) with many arguments for customization, and extensions possible. Uses a global options() system for neatly reducing redundant code. Also contains tools for immediate saving of objects and returning a hashed link to the object, useful for creating download links to high resolution images upon rendering in 'Quarto'. Suitable for highly customized reports, primarily intended for survey research. Complements https://CRAN.R-project.org/package=saros.base.

See also

Author

Maintainer: Stephan Daus stephus.daus@gmail.com (ORCID) [copyright holder]

Other contributors:

Julia Silge (Author of internal scale_x_reordered) [contributor]

David Robinson (Author of internal scale_x_reordered) [contributor]

Nordic Institute for The Studies of Innovation, Research and Education (NIFU) [funder]

Kristiania University College [funder]